requirements.txt
# Root: deps cho scripts (meta_*, v.v.). Core app dùng ai_content_director/requirements.txt.
# Đã bỏ: rq (chỉ phục vụ workers/ đã xóa).
python-dotenv==1.0.0
requests==2.31.0
